input_x/a.i/legacy/AIPass_Core/requirements.txt
openai==1.82.0
anthropic==0.25.0
mistralai==0.2.0
google-generativeai==0.3.2
langchain==0.3.25
tiktoken==0.9.0
pytest==8.3.5
requests==2.32.3
